+++ source/val/validate_adjacency.cpp
@@ -15,14 +15,14 @@
 // Validates correctness of the intra-block preconditions of SPIR-V
 // instructions.
 
-#include "validate.h"
+#include "source/val/validate.h"
 
 #include <string>
 
-#include "diagnostic.h"
-#include "opcode.h"
-#include "val/instruction.h"
-#include "val/validation_state.h"
+#include "source/diagnostic.h"
+#include "source/opcode.h"
+#include "source/val/instruction.h"
+#include "source/val/validation_state.h"
 
 namespace spvtools {
 namespace val {
