+++ source/val/validate_capability.cpp
@@ -14,15 +14,15 @@
 
 // Validates OpCapability instruction.
 
-#include "validate.h"
+#include "source/val/validate.h"
 
 #include <cassert>
 #include <unordered_set>
 
-#include "diagnostic.h"
-#include "opcode.h"
-#include "val/instruction.h"
-#include "val/validation_state.h"
+#include "source/diagnostic.h"
+#include "source/opcode.h"
+#include "source/val/instruction.h"
+#include "source/val/validation_state.h"
 
 namespace spvtools {
 namespace val {
