+++ source/val/validate_conversion.cpp
@@ -14,12 +14,12 @@
 
 // Validates correctness of conversion instructions.
 
-#include "validate.h"
+#include "source/val/validate.h"
 
-#include "diagnostic.h"
-#include "opcode.h"
-#include "val/instruction.h"
-#include "val/validation_state.h"
+#include "source/diagnostic.h"
+#include "source/opcode.h"
+#include "source/val/instruction.h"
+#include "source/val/validation_state.h"
 
 namespace spvtools {
 namespace val {
