+++ source/val/validate_datarules.cpp
@@ -14,17 +14,17 @@
 
 // Ensures Data Rules are followed according to the specifications.
 
-#include "validate.h"
+#include "source/val/validate.h"
 
 #include <cassert>
 #include <sstream>
 #include <string>
 
-#include "diagnostic.h"
-#include "opcode.h"
-#include "operand.h"
-#include "val/instruction.h"
-#include "val/validation_state.h"
+#include "source/diagnostic.h"
+#include "source/opcode.h"
+#include "source/operand.h"
+#include "source/val/instruction.h"
+#include "source/val/validation_state.h"
 
 namespace spvtools {
 namespace val {
