+++ source/val/validate_literals.cpp
@@ -14,14 +14,14 @@
 
 // Validates literal numbers.
 
-#include "validate.h"
+#include "source/val/validate.h"
 
 #include <cassert>
 
-#include "diagnostic.h"
-#include "opcode.h"
-#include "val/instruction.h"
-#include "val/validation_state.h"
+#include "source/diagnostic.h"
+#include "source/opcode.h"
+#include "source/val/instruction.h"
+#include "source/val/validation_state.h"
 
 namespace spvtools {
 namespace val {
